requirements.txt
typer==0.12.3
click==8.1.7
python-binance==1.0.19
python-dotenv==1.0.1